Basic Weather Carousel With Inital Styling
