--- src/components/weather-page/SingleDaySummary.scss
@@ -1,69 +1,0 @@
-@mixin center-in-container {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);    
-}
-
-
-
-body {
-    background-color: black;
-}
-
-.SingleDaySummary {
-
-    position: relative;
-    margin: 20px;
-    color: white;
-
-    width: 360px;
-    height: 360px;
-    background-color: blueviolet;
-
-
-    .DayIconTempCluster {
-        position: absolute;
-        width: 50%;
-        margin: 40px;
-        background-color: blue;
-
-        svg {
-            position: absolute;
-            width: 40px;
-            height: 40px;
-            fill: white;
-        }
-    
-        .TempContainer {
-            position: absolute;    
-            font-weight: bold;        
-            font-size: 24px;
-            width: 80px;
-            height: 80px;           
-        }
-
-    }
-
-    .TextSummaryContainer {
-
-        position: absolute;
-        width: 50%;
-        top: 50%;
-        left: 50%;
-
-        .SummaryText {
-
-            @include center-in-container;        
-            word-wrap: normal;  
-            font-size: 14px;
-            padding: 10%;
-            background-color: red;
-        }
-
-    }
-    
-
-    
-
-}
