+++ src/components/weather-page/singleDay.scss
@@ -1,0 +1,92 @@
+body {
+    background-color: black;
+}
+
+.SingleDaySummary {
+
+    margin: 2px;
+
+    background-color: rgba(230, 230, 230, 0.1);   
+    background: linear-gradient(90deg, rgba(230,230,230,0.1) 0%, rgba(255,255,255,0.15) 100%, rgba(220,220,220,0.1) 100%);
+    border-style: solid;
+    border-width: 1px;
+    border-color:  rgba(230, 230, 230, 0.5);
+    width: 60px;
+    height: 240px;        
+    display: flex;
+    flex-direction: column;
+    align-content: start;
+
+    color: rgba(255,255,255,0.7);
+
+
+    .selectionTab {
+        margin-top: -2px;
+        width: 60px;
+        height: 4px;        
+    }
+
+    .weekday {   
+        height: 160px;
+        width: 100%;
+        margin-top: 20px;
+        text-transform: uppercase; 
+        font-weight: bold;   
+        font-size: 18px;
+        letter-spacing: -4px;
+        
+        text-orientation: upright;
+        writing-mode: vertical-lr;         
+    }
+
+    .icon {           
+        height: 40px;
+        width: 100%;
+        margin-left: 15px;    
+
+        svg {
+            width: 30px;
+            height: 30px;
+            fill: white;
+        }   
+    }
+
+    .AverageTemp {
+        
+        height: 40px;
+        width: 50%;
+        
+        margin-left: 15px;
+
+        font-weight: bold;   
+        font-size: 18px;
+        color: white;
+    }   
+}
+
+.SingleDaySummary:hover {
+    background-color: rgba(230, 230, 230, 0.2);  
+    background: linear-gradient(90deg, rgba(230,230,230,0.2) 0%, rgba(255,255,255,0.3) 100%, rgba(220,220,220,0.2) 100%);
+    color: rgba(255,255,255,1); 
+    .selectionTab {
+        background-color: red;
+    }
+};
+
+.SingleDayDetail {
+    margin: 2px;
+
+    background-color: rgba(230, 230, 230, 0.1);   
+    background: linear-gradient(90deg, rgba(230,230,230,0.1) 0%, rgba(255,255,255,0.15) 100%, rgba(220,220,220,0.1) 100%);
+    border-style: solid;
+    border-width: 1px;
+    border-color:  rgba(230, 230, 230, 0.5);
+    width: 240px;
+    height: 240px;  
+
+    color: rgba(255,255,255,0.7);
+
+    padding-left: 10%;
+    padding-right: 10%;
+    
+}
