+++ src/components/weather-page/weather.scss
@@ -1,0 +1,23 @@
+
+.weather-page-title {
+    margin: 60px;
+    color: white;  
+}
+
+.weather-page-seven-day-carousel {
+
+    margin: 60px;
+
+    display: flex;
+    flex-direction: row;
+
+    border-style: solid;
+    border-width: 1px;
+    border-color:  rgba(230, 230, 230, 0.1);
+
+    background-color: rgba(230, 230, 230, 0.05); 
+
+    
+    
+
+}
